Backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
flask-mail>=0.9.1
python-dotenv>=1.0.0
